requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
mangum>=0.17.0
httpx>=0.27.0
tenacity>=8.3.0
